requirements.txt
pymupdf4llm
ebooklib
html2text
qdrant-client
openai
fastapi
uvicorn
jinja2
PyYAML
python-slugify
python-dotenv
